backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pymysql==1.1.1
cryptography==44.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
groq>=0.9.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
